Fixed failing CI checks.
